Convert vue app to use single file components
