+++ package.json
@@ -245,7 +245,7 @@
         "compile": "webpack",
         "compile:tsc": "tsc -p ./",
         "postcompile": "npm run webview",
-        "watch": "tsc-watch -p ./ --onSuccess \"npm run webview\"",
+        "watch": "tsc-watch -p ./",
         "postinstall": "node ./node_modules/vscode/bin/install",
         "test": "npm run compile:tsc && node ./node_modules/vscode/bin/test",
         "webview": "cd webview && npm run build",
