+++ webview/src/components/FormSection.vue
@@ -1,0 +1,40 @@
+<template>
+    <div class="form-section">
+        <div>
+            <message-input
+                v-bind:onSubmit="onSubmit"
+                v-bind:placeholder="placeholder">
+            </message-input>
+            <status-text v-bind:status="status"></status-text>
+        </div>
+    </div>
+</template>
+
+<script>
+import Vue from 'vue';
+import StatusText from './StatusText.vue'
+import MessageInput from './MessageInput.vue'
+import { sendMessage } from '../utils';
+
+export default {
+    name: 'form-section',
+    props: ["channel", "status", "atMentions"],
+    components: {
+        StatusText, MessageInput
+    },
+    computed: {
+        placeholder: function() {
+            return `Message ${this.channel ? this.channel.name : ""}`;
+        }
+    },
+    methods: {
+        onSubmit: function(text) {
+            const type = text.startsWith("/") ? "command" : "text";
+            sendMessage(text, type);
+        }
+    },
+    mounted() {
+        return sendMessage("is_ready", "internal");
+    }
+}
+</script>
