+++ webview/src/components/MessageAuthor.vue
@@ -1,0 +1,13 @@
+<template>
+    <div class="msg-author">
+        <img v-bind:src="content.authorIcon" />
+        <span>{{ content.author }}</span>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'message-author',
+    props: ['content']
+}
+</script>
