+++ webview/src/components/MessageReaction.vue
@@ -1,0 +1,13 @@
+<template>
+    <li>
+        <div>{{emoji}}</div>
+        <div>{{count}}</div>
+    </li>
+</template>
+
+<script>
+export default {
+    name: 'message-reaction',
+    props: ['emoji', 'count', 'users']
+}
+</script>
