+++ webview/src/components/StatusText.vue
@@ -1,0 +1,10 @@
+<template>
+    <div class="status-text">{{ status }}</div>
+</template>
+
+<script>
+export default {
+    name: 'status-text',
+    props: ['status']
+}
+</script>
