+++ webview/src/main.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import App from './App.vue'
 
 var app = new Vue({
-    template: `<app-container :messages="messages" :users="users" :channel="channel"> </app-container>`,
+    template: `<app :messages="messages" :users="users" :channel="channel"> </app>`,
     // TODO: add fontFamily and fontSize
     props: ['messages', 'users', 'channel', 'statusText', 'atMentions'],
     components: { App },
@@ -10,9 +10,10 @@ var app = new Vue({
 })
 
 window.addEventListener('message', message => {
+    // TODO: we don't need to de-structure here, can do inside app-container
     app.messages = message.data.messages;
     app.users = message.data.users;
-    app.channel = message.data.channel
-    app.statusText = message.data.statusText
-    app.atMentions = message.data.atMentions
+    app.channel = message.data.channel;
+    app.statusText = message.data.statusText;
+    app.atMentions = message.data.atMentions;
 })
